Integration tests setup fixes (Git, CoilMQ)
Git setup wasn't working as expected, leaving the respository in
detached HEAD status. Using git.cmd.Git instead of the repository
object works fine. CoilMQ is not closing listening socket on shutdown,
so if we try to use the same socket again, it won't work. Just closing
the socket manually to fix the issues.