Review YAML deserialization for Ruby symbols
Right now, when we get a Ruby symbol (!ruby/symbol tag) as part of a
YAML message this is treated as a regular string (no leading colon),
however this leads to errors since Ruby symbols can be sent as regular
string (just with leading colon).

In order to avoid this situation !ruby/symbol tag will be translated
into string with leading colon, so we get the same result always.

This change also require some updates on the SSL security provider.